requirements.txt
openai
transformers
huggingface_hub
python-docx
pdfplumber
tk
PyMuPDF